requirements.txt
langchain
langchain-core
python-dotenv
langchain-groq
streamlit
langchain-openai
faiss-cpu
langchain-community
pypdf
